requirements.txt
langchain
langchain-nvidia-ai-endpoints
python-dotenv
